A few models done
Could be more. I'm trying my best to spend more time working on a project but it's not always enough.
No methods were added by now. This is awaiting in the future.
The fields are not complete so trying to use them right now will probably not be fully possible.