Selenium is done but without coverage
For it to run correctly you need to have your webdriver in test/selenium_dir/selenium_webdrivers and it's name in test/../settings.py (beginning of the file).
The code for the selenium tests is terrible but I decided that there's no time to make a proper structure. I do things too slowly even without it.